fix: no animation in nested stacks (#1012)
PR fixing navigating in nested stack without animation, which did not trigger navigation events due to check for being nested, resolving in never dispatching those events there. PR switches it to a check inside parent fragment and makes it decide whether it should dispatch navigation events in a child.
We don't do it if the child has none animation and we are going forward since events will be dispatched in child via onCreateAnimation of ScreenStackFragment then.